api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
